+++ src/search/pdbs/pdb_factory.cc
@@ -26,7 +26,7 @@ static options::PluginTypePlugin<PDBFactory> _type_plugin(
     PDBFactory::compute_pdb(const TaskProxy & task, 
 			   const Pattern &pattern, 
 			    const std::vector<int> &operator_costs, 
-			    int time_limit) {
+			    double time_limit) {
 	assert(!pattern.empty ());
 	num_patterns_requested ++;
 	auto item = stored_pdbs.find(PDBKey(pattern, operator_costs));
@@ -40,7 +40,6 @@ static options::PluginTypePlugin<PDBFactory> _type_plugin(
 	}
 
 	num_patterns_created ++;
-        //cout<<"compute_before_create,g_timer:"<<utils::g_timer()<<endl;
 	shared_ptr<PatternDatabaseInterface> result = create_pdb(task, pattern, operator_costs, time_limit); 
         //cout<<"compute_after create,g_timer:"<<utils::g_timer()<<endl;
 	stored_pdbs[PDBKey(pattern, operator_costs)] = result;
